+++ lib/slack-ruby-bot/mvc/mvc.rb
@@ -1,0 +1,5 @@
+require_relative '../mvc/controller/base'
+
+require_relative '../mvc/model/base'
+
+require_relative '../mvc/view/base'