+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.10.2 (Next)
 
 * Your contribution here.
+* [#137](https://github.com/slack-ruby/slack-ruby-bot/pull/137): Add Model-View-Controller classes to allow for more explicit control over how `command`s are designed - [@chuckremes](https://github.com/chuckremes).
 * [#130](https://github.com/slack-ruby/slack-ruby-bot/issues/130): Added test dependencies in TUTORIAL.md - [@jbristow](https://github.com/jbristow).
 
 ### 0.10.1 (2/12/2017)